Adjust column ratio for bottom half of nav bar
Initially, the two rows in the grid that made up the nav bar conformed
to the same template row ratio, however I wanted that to differ for the
elements in the bottom half of the nav bar.
Thus, within the nav bar grid, I created two further grids which
encapsulate the top half and bottom half respectively.